s3/requirements.txt
pdfplumber==0.10.3
docx2txt==0.8
spacy==3.7.2
google-generativeai==0.3.2
